requirements.txt
mysql-connector-python
Flask
openai
requests
authlib
python-dotenv
